Reorganizing gemviz23 folder
